fixed namespacing, erased redundant test cases, and also fixed more small issues with previous commit
