+++ Fakebook.Posts/Fakebook.Posts.UnitTests/PostTest.cs
@@ -5,12 +5,13 @@ using System.Text;
 using System.Threading.Tasks;
 using Fakebook.Posts.Domain.Models;
 using Xunit;
+
 namespace Fakebook.Posts.UnitTests
 {
-
     public class PostTests
     {
         private const int Id = 1;
+
         [Fact]
         public void Post_Constructor_Pass()
         {
@@ -24,7 +25,6 @@ namespace Fakebook.Posts.UnitTests
             Assert.NotNull(newpost);
         }
 
-
         [Fact]
         public void Post_EmailIsNotNull()
         {
@@ -32,27 +32,24 @@ namespace Fakebook.Posts.UnitTests
             const string Content = "This is some content";
             const string userEmail = "damion.silver@test.com";
             //act
-            Post newcomment = new Post(userEmail, Content);
+            Post newpost = new Post(userEmail, Content);
 
             //assert
-            Assert.NotNull(newcomment.UserEmail);
+            Assert.NotNull(newpost.UserEmail);
         }
-
 
         [Fact]
         public void Post_EmailEquality()
         {
-
             //arrange
             const string Content = "This is some content";
             const string userEmail = "damion.silver@test.com";
             //act
-            Post newcomment = new Post(userEmail, Content);
+            Post newpost = new Post(userEmail, Content);
 
             //assert
-            Assert.Equal(userEmail, newcomment.UserEmail);
+            Assert.Equal(userEmail, newpost.UserEmail);
         }
-
 
         [Fact]
         public void Post_ContentIsNotNull()
@@ -61,10 +58,10 @@ namespace Fakebook.Posts.UnitTests
             const string Content = "This is some content";
             const string userEmail = "damion.silver@test.com";
             //act
-            Post newcomment = new Post(userEmail, Content);
+            Post newpost = new Post(userEmail, Content);
 
             //assert
-            Assert.NotNull(newcomment.Content);
+            Assert.NotNull(newpost.Content);
         }
 
         [Fact]
@@ -74,21 +71,67 @@ namespace Fakebook.Posts.UnitTests
             const string Content = "This is some content";
             const string userEmail = "damion.silver@test.com";
             //act
-            Post newcomment = new Post(userEmail, Content);
+            Post newpost = new Post(userEmail, Content);
 
             //assert
-            Assert.Contains("@", newcomment.UserEmail);
+            Assert.Contains("@", newpost.UserEmail);
         }
 
         [Fact]
         public void Post_EmailFormationThrowsException()
         {
-
             const string Content = "This is some content";
             const string userEmail = "damion.silvertest.com";
 
             Assert.Throws<ArgumentException>(() => new Post(userEmail, Content));
+        }
 
+        [Fact]
+        public void Comment_Constructor_Pass()
+        {
+            //arrange
+            const string Content = "This is some content";
+            const string userEmail = "damion.silver@test.com";
+            //act
+            Comment newcomment = new Comment(userEmail, Content);
+
+            //assert
+            Assert.NotNull(newcomment);
+        }
+
+        [Fact]
+        public void Comment_EmailEquality()
+        {
+            //arrange
+            const string Content = "This is some content";
+            const string userEmail = "damion.silver@test.com";
+            //act
+            Comment newcomment = new Comment(userEmail, Content);
+
+            //assert
+            Assert.Equal(userEmail, newcomment.UserEmail);
+        }
+
+        [Fact]
+        public void Comment_ContentIsNotNull()
+        {
+            //arrange
+            const string Content = "This is some content";
+            const string userEmail = "damion.silver@test.com";
+            //act
+            Comment newcomment = new Comment(userEmail, Content);
+
+            //assert
+            Assert.NotNull(newcomment.Content);
+        }
+
+        [Fact]
+        public void Comment_EmailFormationThrowsException()
+        {
+            const string Content = "This is some content";
+            const string userEmail = "damion.silvertest.com";
+
+            Assert.ThrowsAny<ArgumentException>(() => new Comment(userEmail, Content));
         }
     }
 }
